:white_check_mark: Fix unit test, move auth enpoints from app to auth controller
