+++ src/users/schemas/user.schema.ts
@@ -1,12 +1,15 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document, ObjectId } from 'mongoose';
+import { Document, SchemaTypes, Types } from 'mongoose';
 
 export type UserDocument = User & Document;
 
 @Schema()
 export class User {
-  @Prop()
-  _id: number;
+  @Prop({
+    type: SchemaTypes.ObjectId,
+    ref: User.name,
+  })
+  _id: Types.ObjectId;
 
   @Prop()
   name: string;
